feat(refactor): Move InMemoryDecisionAdapter to its own file and add JSDoc
   - Moved InMemoryDecisionAdapter from src/decisionAdapter.ts to src/InMemoryDecisionAdapter.ts for better module organization.
   - Deleted the now-empty src/decisionAdapter.ts file.
   - Added JSDoc comments to src/localStorageDecisionAdapter.ts for improved documentation.
